Clean code & Move some fuzzer functions into `utils.rs`
