+++ cairo-native-fuzzer/src/fuzzer/fuzzer.rs
@@ -2,147 +2,42 @@ use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use std::time::Instant;
 
-use crate::runner::runner::compile_sierra_program;
-use crate::runner::runner::create_executor;
 use cairo_lang_compiler::CompilerConfig;
 use cairo_lang_sierra::ids::FunctionId;
 use cairo_lang_sierra::program::Program;
 use cairo_lang_starknet::compile::compile_path;
 use cairo_native::context::NativeContext;
-use colored::*;
+use cairo_native::executor::JitNativeExecutor;
 use starknet_types_core::felt::Felt;
 
 use crate::fuzzer::statistics::FuzzerStats;
-use crate::mutator::argument_type::{map_argument_type, ArgumentType};
+use crate::fuzzer::utils::{
+    find_entry_point_id, get_function_argument_types, print_contract_functions, print_init_message,
+};
+use crate::mutator::argument_type::ArgumentType;
 use crate::mutator::basic_mutator::Mutator;
-use crate::runner::runner::run_program;
-use crate::utils::get_function_by_id;
+use crate::runner::runner::{compile_sierra_program, create_executor, run_program};
 
-#[warn(dead_code)]
+/// Struct representing the fuzzer
 pub struct Fuzzer {
+    // Path to the Cairo program
     program_path: PathBuf,
+    // Entry point of the Sierra program
     entry_point: Option<String>,
+    // Sierra program
     sierra_program: Option<Arc<Program>>,
+    // Entry point parameters
     params: Arc<Mutex<Vec<Felt>>>,
+    // ID of the entry point
     entry_point_id: Option<FunctionId>,
+    // Mutator for the parameters
     mutator: Option<Mutex<Mutator>>,
+    // Types of the entry point arguments
     argument_types: Vec<ArgumentType>,
+    // Fuzzer statistics
     stats: Arc<Mutex<FuzzerStats>>,
+    // Native context
     native_context: NativeContext,
-}
-
-/// Print the initial message with the seed
-fn print_init_message(seed: u64) {
-    println!(
-        "
-=============================================================================================================================================================
-╔═╗ ┌─┐ ┬ ┬─┐ ┌───┐   ╔═╗ ┬ ┬ ┌─┐ ┌─┐ ┌─┐ ┬─┐      | Seed -- {}
-║   ├─┤ │ ├┬┘ │2.0│───╠╣  │ │ ┌─┘ ┌─┘ ├┤  ├┬┘      |
-╚═╝ ┴ ┴ ┴ ┴└─ └───┘   ╚   └─┘ └─┘ └─┘ └─┘ ┴└─      |
-=============================================================================================================================================================\n",
-        seed,
-    );
-}
-
-/// Print the contract functions prototypes
-fn print_contract_functions(sierra_program: &Option<Arc<Program>>) {
-    println!("Contract functions :\n");
-
-    if let Some(program) = sierra_program {
-        for function in &program.funcs {
-            let function_name = function
-                .id
-                .debug_name
-                .as_ref()
-                .map_or_else(|| "unknown".to_string(), |name| name.to_string());
-
-            let signature = &function.signature;
-
-            // Collect parameter types
-            let param_types: Vec<String> = signature
-                .param_types
-                .iter()
-                .map(|param| {
-                    param
-                        .debug_name
-                        .as_ref()
-                        .map_or_else(|| "unknown".to_string(), |name| name.to_string())
-                })
-                .collect();
-
-            // Collect return types
-            let ret_types: Vec<String> = signature
-                .ret_types
-                .iter()
-                .map(|ret_type| {
-                    ret_type
-                        .debug_name
-                        .as_ref()
-                        .map_or_else(|| "unknown".to_string(), |name| name.to_string())
-                })
-                .collect();
-
-            // Format the prototype
-            let prototype = format!(
-                "{} ({}) -> ({})",
-                function_name.bold().white(),
-                param_types.join(", ").green(),
-                ret_types.join(", ").cyan()
-            );
-
-            // Print the contract functions
-            println!("- {}", prototype);
-        }
-    }
-}
-
-/// Find the entry point id
-fn find_entry_point_id(sierra_program: &Option<Arc<Program>>, entry_point: &str) -> FunctionId {
-    let sierra_program = sierra_program
-        .as_ref()
-        .expect("Sierra program not available");
-    cairo_native::utils::find_function_id(sierra_program, entry_point)
-        .expect(&format!("Entry point '{}' not found", entry_point))
-        .clone()
-}
-
-/// Returns a vector of the function parameter types
-///
-/// For example, given a function with the prototype:
-/// ```
-/// myfunction(a: felt252, b: felt252) -> felt252
-/// ```
-/// This function will return:
-/// ```
-/// [Felt, Felt]
-/// ```
-fn get_function_argument_types(
-    sierra_program: &Option<Arc<Program>>,
-    entry_point_id: &Option<FunctionId>,
-) -> Vec<ArgumentType> {
-    let func = match (sierra_program, entry_point_id) {
-        (Some(program), Some(entry_point_id)) => get_function_by_id(program, entry_point_id),
-        _ => None,
-    };
-
-    if let Some(func) = func {
-        let argument_types: Vec<ArgumentType> = func
-            .signature
-            .param_types
-            .iter()
-            .filter_map(|param_type| {
-                if let Some(debug_name) = &param_type.debug_name {
-                    // Map param_type to an `ArgumentType`
-                    // For now we only handle felt252
-                    return map_argument_type(debug_name);
-                }
-                None
-            })
-            .collect();
-        argument_types
-    } else {
-        Vec::new()
-    }
 }
 
 impl Fuzzer {
@@ -163,14 +58,13 @@ impl Fuzzer {
         }
     }
 
-    /// Init the fuzzer with a given seed
-    /// - Initialize the mutator with the given seed
-    /// - Compile Cairo code to Sierra
-    /// - Find the entry id
+    /// Initialize the fuzzer with a given seed
+    /// - Initializes the mutator with the given seed
+    /// - Compiles Cairo code to Sierra
+    /// - Finds the entry point ID
     pub fn init(&mut self, seed: u64) -> Result<(), String> {
         print_init_message(seed);
 
-        println!("[+] Initializing mutator with seed: {}", seed);
         self.mutator = Some(Mutex::new(Mutator::new(seed)));
 
         println!("[+] Compiling Cairo contract to Sierra");
@@ -206,7 +100,7 @@ impl Fuzzer {
         Ok(())
     }
 
-    /// Generate params based on the function argument types
+    /// Generates parameters based on the function argument types.
     pub fn generate_params(&mut self) {
         let mut params = self.params.lock().unwrap();
         *params = self
@@ -219,11 +113,80 @@ impl Fuzzer {
             .collect();
     }
 
-    /// Run the fuzzer
-    pub fn fuzz(&mut self, iter: i32) -> Result<(), String> {
+    /// Initializes parameters based on the function argument types.
+    fn initialize_parameters(&mut self) {
         self.argument_types =
             get_function_argument_types(&self.sierra_program, &self.entry_point_id);
         self.generate_params();
+    }
+
+    /// Sets up the execution environment by compiling the Sierra program and creating a JIT executor.
+    fn setup_execution_environment(&self) -> Result<JitNativeExecutor, String> {
+        let mlir_module = compile_sierra_program(
+            &self.native_context,
+            self.sierra_program
+                .as_ref()
+                .ok_or("Sierra program not available")?,
+        )?;
+        let executor = create_executor(mlir_module);
+        Ok(executor)
+    }
+
+    /// Executes the program and checks for crashes.
+    fn execute_program(&self, executor: &JitNativeExecutor) -> Result<bool, String> {
+        let params_guard = self.params.lock().unwrap();
+        match run_program(
+            executor,
+            self.entry_point_id.as_ref().unwrap(),
+            &params_guard,
+        ) {
+            Ok(result) => {
+                if result.failure_flag {
+                    println!("Results : {:?}", result);
+                    println!("Crash detected! Exiting fuzzer.");
+                    return Ok(true);
+                }
+            }
+            Err(e) => eprintln!("Error during execution: {}", e),
+        }
+        Ok(false)
+    }
+
+    /// Mutates the parameters using the mutator.
+    fn mutate_parameters(&self) {
+        let mut mutator_guard = self.mutator.as_ref().unwrap().lock().unwrap();
+        for param in self.params.lock().unwrap().iter_mut() {
+            *param = mutator_guard.mutate(*param);
+        }
+    }
+
+    /// Prints the statistics every 1000 executions.
+    fn print_statistics(&self, current_iter: i32) {
+        if current_iter % 1000 == 0 && current_iter != 0 {
+            let stats_guard = self.stats.lock().unwrap();
+            let uptime = stats_guard.start_time.elapsed();
+            let uptime_secs = uptime.as_secs_f64();
+
+            // Calculate execs per second
+            let execs_per_second = if uptime_secs > 0.0 {
+                stats_guard.total_executions as f64 / uptime_secs
+            } else {
+                0.0
+            };
+
+            println!(
+                "| {:<30} | {:<25} | {:<25} | {:<20} |",
+                format!("Total Executions = {}", stats_guard.total_executions),
+                format!("Uptime = {:.1}s", uptime_secs),
+                format!("Crashes = {}", stats_guard.crashes),
+                format!("Exec Speed = {:.2} execs/s", execs_per_second)
+            );
+        }
+    }
+
+    /// Runs the fuzzer.
+    pub fn fuzz(&mut self, iter: i32) -> Result<(), String> {
+        self.initialize_parameters();
 
         // Initialize the start time
         {
@@ -234,51 +197,23 @@ impl Fuzzer {
         let mut current_iter = 0;
         let max_iter = if iter == -1 { i32::MAX } else { iter };
 
-        // Compile Sierra program into a MLIR module
-        let mlir_module = compile_sierra_program(
-            &self.native_context,
-            self.sierra_program
-                .as_ref()
-                .ok_or("Sierra program not available")?,
-        )?;
+        let executor = self.setup_execution_environment()?;
 
-        // Create a JIT executor
-        let executor = create_executor(mlir_module);
-
-        // Infinite loop of execution
+        // Main fuzz loop
         loop {
             if current_iter >= max_iter {
                 println!("Maximum iterations reached. Exiting fuzzer.");
                 break;
             }
 
-            let params_guard = self.params.lock().unwrap();
-
-            // Execute the program
-            match run_program(
-                &executor,
-                self.entry_point_id.as_ref().unwrap(),
-                &params_guard,
-            ) {
-                Ok(result) => {
-                    if result.failure_flag {
-                        println!("Results : {:?}", result);
-                        println!("Crash detected! Exiting fuzzer.");
-
-                        // Increment the crashes counter
-                        {
-                            let mut stats_guard = self.stats.lock().unwrap();
-                            stats_guard.crashes += 1;
-                        }
-
-                        break;
-                    }
+            if self.execute_program(&executor)? {
+                // Increment the crashes counter
+                {
+                    let mut stats_guard = self.stats.lock().unwrap();
+                    stats_guard.crashes += 1;
                 }
-                Err(e) => eprintln!("Error during execution: {}", e),
+                break;
             }
-
-            // Release the lock before mutating params
-            drop(params_guard);
 
             // Increment the total_executions counter
             {
@@ -286,33 +221,8 @@ impl Fuzzer {
                 stats_guard.total_executions += 1;
             }
 
-            // Mutate params using the mutator
-            let mut mutator_guard = self.mutator.as_ref().unwrap().lock().unwrap();
-            for param in self.params.lock().unwrap().iter_mut() {
-                *param = mutator_guard.mutate(*param);
-            }
-
-            // Print stats every 1000 executions
-            if current_iter % 1000 == 0 && current_iter != 0 {
-                let stats_guard = self.stats.lock().unwrap();
-                let uptime = stats_guard.start_time.elapsed();
-                let uptime_secs = uptime.as_secs_f64();
-
-                // Calculate execs per second
-                let execs_per_second = if uptime_secs > 0.0 {
-                    stats_guard.total_executions as f64 / uptime_secs
-                } else {
-                    0.0
-                };
-
-                println!(
-                    "| {:<30} | {:<25} | {:<25} | {:<20} |",
-                    format!("Total Executions = {}", stats_guard.total_executions),
-                    format!("Uptime = {:.1}s", uptime_secs),
-                    format!("Crashes = {}", stats_guard.crashes),
-                    format!("Exec Speed = {:.2} execs/s", execs_per_second)
-                );
-            }
+            self.mutate_parameters();
+            self.print_statistics(current_iter);
 
             current_iter += 1;
         }
