+++ cairo-native-fuzzer/src/fuzzer/mod.rs
@@ -1,2 +1,3 @@
 pub mod fuzzer;
 pub mod statistics;
+pub mod utils;
